the translation of Nat.le elimates proof to create non-proofs. bad.
(*
Debug:
(fix
 Coq_Init_Peano_le_RR0 (n n₂ : nat) (n_R : Coq_Init_Datatypes_nat_RR0 n
n₂)
                       (H H0 : nat) (H1 : Coq_Init_Datatypes_nat_RR0 H
H0)
                       (H2 : (n <= H)%nat) (H3 : (n₂ <= H0)%nat) {struct
H2} :
   Prop :=
   match H2 with
   | le_n _ => match H3 with
               | le_n _ => True
               | le_S _ _ _ => False
               end
   | le_S _ m x =>
       match H3 with
       | le_n _ => False
       | le_S _ m₂ x0 =>
           {m_R : Coq_Init_Datatypes_nat_RR0 m m₂ &
           {_ : Coq_Init_Peano_le_RR0 n n₂ n_R m m₂ m_R x x0 & True}}
       end
   end)
File "./paramDirect.v", line 720, characters 0-64:
Error:
Incorrect elimination of "H3" in the inductive type "le":
the return type has sort "Type" while it should be "Prop".
Elimination of an inductive object of sort Prop
is not allowed on a predicate in sort Type

*)

Inductive props need to be translated as fun _ _ => True? because in the
the new translation we never match on n_R, this would be okay?