documented where having typing info in Opid helps. keep casts in ret type?
Otherwise we will have 2 typing rules, parhaps paramaterized by a bool
enforcing whether casts are present.

If casts are in the terms, not Opid, we will need a proof that
substitution preserves hasSortPropOrSet. This is not true in general and
we would need some property about the kinds of substitution we allow --
type respecting?