too bad the other inductive subgoal is not provable.
[[\x:A.B]] may not be in BC.
[A] may have [x] as its boundvars. x is bound in B, but NOT in A/
[[A]] may have [x] or [x'] or [xr] as bound variables (but not free vars).
In [[\x:A.B]], [[A]] occurs in the scope of [x] and [x'].
Thus it may not in BC.

The Pi case has similar problems.

Nevertheless, BC may be a sufficient condition for [[]] translate
to be well-defined w.r.t alpha equality. I don't have any proof.

one possibility may be to add to the safe pedicate that [A] cannot
mention x in  \x:A.B and \forall x:A.B. Does bcSubst
then preserve the safe predicate?